(tombstone-dev): delete old file and rename new file (#232)
