+++ pkg/vm/engine/disttae/db.go
@@ -255,7 +255,7 @@ func (e *Engine) GetLatestCatalogCache() *cache.CatalogCache {
 func (e *Engine) getOrCreateSnapPart(
 	ctx context.Context,
 	tbl *txnTable,
-	ts types.TS) (*logtailreplay.PartitionStateInProgress, error) {
+	ts types.TS) (*logtailreplay.PartitionState, error) {
 
 	//check whether the latest partition is available for reuse.
 	// if the snapshot-read's ts is too old , subscribing table maybe timeout.
@@ -295,7 +295,7 @@ func (e *Engine) getOrCreateSnapPart(
 	}
 	err = snap.ConsumeSnapCkps(ctx, ckps, func(
 		checkpoint *checkpoint.CheckpointEntry,
-		state *logtailreplay.PartitionStateInProgress) error {
+		state *logtailreplay.PartitionState) error {
 		locs := make([]string, 0)
 		locs = append(locs, checkpoint.GetLocation().String())
 		locs = append(locs, strconv.Itoa(int(checkpoint.GetVersion())))
@@ -397,7 +397,7 @@ func (e *Engine) LazyLoadLatestCkp(
 
 	if err := part.ConsumeCheckpoints(
 		ctx,
-		func(checkpoint string, state *logtailreplay.PartitionStateInProgress) error {
+		func(checkpoint string, state *logtailreplay.PartitionState) error {
 			entries, closeCBs, err := logtail.LoadCheckpointEntries(
 				ctx,
 				e.service,
