+++ pkg/vm/engine/disttae/filter.go
@@ -986,7 +986,7 @@ func TryFastFilterBlocks(
 	snapshotTS timestamp.Timestamp,
 	tableDef *plan.TableDef,
 	exprs []*plan.Expr,
-	snapshot *logtailreplay.PartitionStateInProgress,
+	snapshot *logtailreplay.PartitionState,
 	extraCommittedObjects []objectio.ObjectStats,
 	uncommittedObjects []objectio.ObjectStats,
 	outBlocks *objectio.BlockInfoSlice,
@@ -1027,7 +1027,7 @@ func ExecuteBlockFilter(
 	objectFilterOp ObjectFilterOp,
 	blockFilterOp BlockFilterOp,
 	seekOp SeekFirstBlockOp,
-	snapshot *logtailreplay.PartitionStateInProgress,
+	snapshot *logtailreplay.PartitionState,
 	extraCommittedObjects []objectio.ObjectStats,
 	uncommittedObjects []objectio.ObjectStats,
 	outBlocks *objectio.BlockInfoSlice,
