+++ pkg/vm/engine/disttae/logtail.go
@@ -33,7 +33,7 @@ func consumeEntry(
 	primarySeqnum int,
 	engine *Engine,
 	cache *cache.CatalogCache,
-	state *logtailreplay.PartitionStateInProgress,
+	state *logtailreplay.PartitionState,
 	e *api.Entry,
 ) error {
 	start := time.Now()
@@ -47,7 +47,7 @@ func consumeEntry(
 
 	if state != nil {
 		t0 := time.Now()
-		state.HandleLogtailEntryInProgress(ctx, engine.fs, e, primarySeqnum, packer, engine.mp)
+		state.HandleLogtailEntry(ctx, engine.fs, e, primarySeqnum, packer, engine.mp)
 		v2.LogtailUpdatePartitonConsumeLogtailOneEntryLogtailReplayDurationHistogram.Observe(time.Since(t0).Seconds())
 	}
