+++ pkg/vm/engine/disttae/logtailreplay/partition.go
@@ -30,7 +30,7 @@ import (
 type Partition struct {
 	//lock is used to protect pointer of PartitionState from concurrent mutation
 	lock  chan struct{}
-	state atomic.Pointer[PartitionStateInProgress]
+	state atomic.Pointer[PartitionState]
 
 	// assuming checkpoints will be consumed once
 	checkpointConsumed atomic.Bool
@@ -68,7 +68,7 @@ func NewPartition(
 		lock: lock,
 	}
 	ret.mu.start = types.MaxTs()
-	ret.state.Store(NewPartitionStateInProgress(service, false, id))
+	ret.state.Store(NewPartitionState(service, false, id))
 	return ret
 }
 
@@ -78,7 +78,7 @@ func (r RowID) Less(than RowID) bool {
 	return bytes.Compare(r[:], than[:]) < 0
 }
 
-func (p *Partition) Snapshot() *PartitionStateInProgress {
+func (p *Partition) Snapshot() *PartitionState {
 	return p.state.Load()
 }
 
@@ -86,7 +86,7 @@ func (*Partition) CheckPoint(ctx context.Context, ts timestamp.Timestamp) error 
 	panic("unimplemented")
 }
 
-func (p *Partition) MutateState() (*PartitionStateInProgress, func()) {
+func (p *Partition) MutateState() (*PartitionState, func()) {
 	curState := p.state.Load()
 	state := curState.Copy()
 	return state, func() {
@@ -148,7 +148,7 @@ func (p *Partition) ConsumeSnapCkps(
 	ckps []*checkpoint.CheckpointEntry,
 	fn func(
 		ckp *checkpoint.CheckpointEntry,
-		state *PartitionStateInProgress,
+		state *PartitionState,
 	) error,
 ) (
 	err error,
@@ -200,7 +200,7 @@ func (p *Partition) ConsumeCheckpoints(
 	ctx context.Context,
 	fn func(
 		checkpoint string,
-		state *PartitionStateInProgress,
+		state *PartitionState,
 	) error,
 ) (
 	err error,
