+++ pkg/vm/engine/disttae/logtailreplay/rows_iter_test.go
@@ -156,15 +156,22 @@ func TestPartitionStateRowsIter(t *testing.T) {
 		// delete number i at (deleteAt+i) with (i+1) row id
 		rowIDVec := vector.NewVec(types.T_Rowid.ToType())
 		tsVec := vector.NewVec(types.T_TS.ToType())
+		tombStoneRowIDVec := vector.NewVec(types.T_Rowid.ToType())
+		pkVec := vector.NewVec(types.T_int32.ToType())
+
 		for i := 0; i < num; i++ {
 			vector.AppendFixed(rowIDVec, buildRowID(i+1), false, pool)
 			vector.AppendFixed(tsVec, types.BuildTS(int64(deleteAt+i), 1), false, pool)
+			vector.AppendFixed(tombStoneRowIDVec, types.RandomRowid(), false, pool)
+			vector.AppendFixed(pkVec, int32(i), false, pool)
 		}
 		state.HandleRowsDelete(ctx, &api.Batch{
 			Attrs: []string{"rowid", "time"},
 			Vecs: []api.Vector{
 				mustVectorToProto(rowIDVec),
 				mustVectorToProto(tsVec),
+				mustVectorToProto(pkVec),
+				mustVectorToProto(tombStoneRowIDVec),
 			},
 		}, packer, pool)
 	}
@@ -232,7 +239,12 @@ func TestPartitionStateRowsIter(t *testing.T) {
 		// duplicate delete
 		rowIDVec := vector.NewVec(types.T_Rowid.ToType())
 		tsVec := vector.NewVec(types.T_TS.ToType())
+		tombStoneRowIDVec := vector.NewVec(types.T_Rowid.ToType())
+		pkVec := vector.NewVec(types.T_int32.ToType())
+
 		for i := 0; i < num; i++ {
+			vector.AppendFixed(tombStoneRowIDVec, types.RandomRowid(), false, pool)
+			vector.AppendFixed(pkVec, int32(i), false, pool)
 			vector.AppendFixed(rowIDVec, buildRowID(i+1), false, pool)
 			vector.AppendFixed(tsVec, types.BuildTS(int64(deleteAt+i), 1), false, pool)
 		}
@@ -241,6 +253,8 @@ func TestPartitionStateRowsIter(t *testing.T) {
 			Vecs: []api.Vector{
 				mustVectorToProto(rowIDVec),
 				mustVectorToProto(tsVec),
+				mustVectorToProto(pkVec),
+				mustVectorToProto(tombStoneRowIDVec),
 			},
 		}, packer, pool)
 	}
@@ -303,7 +317,12 @@ func TestInsertAndDeleteAtTheSameTimestamp(t *testing.T) {
 		// delete number i at the same time
 		rowIDVec := vector.NewVec(types.T_Rowid.ToType())
 		tsVec := vector.NewVec(types.T_TS.ToType())
+		tombStoneRowIDVec := vector.NewVec(types.T_Rowid.ToType())
+		pkVec := vector.NewVec(types.T_int32.ToType())
+
 		for i := 0; i < num; i++ {
+			vector.AppendFixed(tombStoneRowIDVec, types.RandomRowid(), false, pool)
+			vector.AppendFixed(pkVec, int32(i), false, pool)
 			vector.AppendFixed(rowIDVec, buildRowID(i+1), false, pool)
 			vector.AppendFixed(tsVec, types.BuildTS(int64(i), 1), false, pool)
 		}
@@ -312,6 +331,8 @@ func TestInsertAndDeleteAtTheSameTimestamp(t *testing.T) {
 			Vecs: []api.Vector{
 				mustVectorToProto(rowIDVec),
 				mustVectorToProto(tsVec),
+				mustVectorToProto(pkVec),
+				mustVectorToProto(tombStoneRowIDVec),
 			},
 		}, packer, pool)
 	}
@@ -375,7 +396,12 @@ func TestDeleteBeforeInsertAtTheSameTime(t *testing.T) {
 		// delete number i at time i with (i+1) row id
 		rowIDVec := vector.NewVec(types.T_Rowid.ToType())
 		tsVec := vector.NewVec(types.T_TS.ToType())
+		tombStoneRowIDVec := vector.NewVec(types.T_Rowid.ToType())
+		pkVec := vector.NewVec(types.T_int32.ToType())
+
 		for i := 0; i < num; i++ {
+			vector.AppendFixed(tombStoneRowIDVec, types.RandomRowid(), false, pool)
+			vector.AppendFixed(pkVec, int32(i), false, pool)
 			vector.AppendFixed(rowIDVec, buildRowID(i+1), false, pool)
 			vector.AppendFixed(tsVec, types.BuildTS(int64(i), 1), false, pool)
 		}
@@ -384,6 +410,8 @@ func TestDeleteBeforeInsertAtTheSameTime(t *testing.T) {
 			Vecs: []api.Vector{
 				mustVectorToProto(rowIDVec),
 				mustVectorToProto(tsVec),
+				mustVectorToProto(pkVec),
+				mustVectorToProto(tombStoneRowIDVec),
 			},
 		}, packer, pool)
 	}
@@ -467,11 +495,14 @@ func TestPrimaryKeyModifiedWithDeleteOnly(t *testing.T) {
 		// delete number i at time i with (i+1) row id
 		rowIDVec := vector.NewVec(types.T_Rowid.ToType())
 		tsVec := vector.NewVec(types.T_TS.ToType())
+		tombStoneRowIDVec := vector.NewVec(types.T_Rowid.ToType())
+
 		primaryKeyVec := vector.NewVec(types.T_int64.ToType())
 		for i := 0; i < num; i++ {
 			vector.AppendFixed(rowIDVec, buildRowID(i+1), false, pool)
 			vector.AppendFixed(tsVec, types.BuildTS(int64(i), 1), false, pool)
 			vector.AppendFixed(primaryKeyVec, int64(i), false, pool)
+			vector.AppendFixed(tombStoneRowIDVec, types.RandomRowid(), false, pool)
 		}
 		state.HandleRowsDelete(ctx, &api.Batch{
 			Attrs: []string{"rowid", "time", "i"},
@@ -479,6 +510,7 @@ func TestPrimaryKeyModifiedWithDeleteOnly(t *testing.T) {
 				mustVectorToProto(rowIDVec),
 				mustVectorToProto(tsVec),
 				mustVectorToProto(primaryKeyVec), // with primary key
+				mustVectorToProto(tombStoneRowIDVec),
 			},
 		}, packer, pool)
 	}
