+++ pkg/vm/engine/disttae/logtailreplay/type.go
@@ -1,0 +1,160 @@
+package logtailreplay
+
+import (
+	"bytes"
+	"fmt"
+	"sync"
+
+	"github.com/matrixorigin/matrixone/pkg/container/batch"
+	"github.com/matrixorigin/matrixone/pkg/container/types"
+	"github.com/matrixorigin/matrixone/pkg/objectio"
+)
+
+type ObjectInfo struct {
+	objectio.ObjectStats
+
+	Appendable  bool
+	Sorted      bool
+	HasDeltaLoc bool
+	CommitTS    types.TS
+	CreateTime  types.TS
+	DeleteTime  types.TS
+}
+
+func (o ObjectInfo) String() string {
+	return fmt.Sprintf(
+		"%s; appendable: %v; sorted: %v; commitTS: %s; createTS: %s; deleteTS: %s",
+		o.ObjectStats.String(), o.Appendable, o.Sorted, o.CommitTS.ToString(),
+		o.CreateTime.ToString(), o.DeleteTime.ToString())
+}
+
+func (o ObjectInfo) Location() objectio.Location {
+	return o.ObjectLocation()
+}
+
+type ObjectEntry struct {
+	ObjectInfo
+}
+
+func (o ObjectEntry) Less(than ObjectEntry) bool {
+	return bytes.Compare((*o.ObjectShortName())[:], (*than.ObjectShortName())[:]) < 0
+}
+
+func (o ObjectEntry) IsEmpty() bool {
+	return o.Size() == 0
+}
+
+func (o *ObjectEntry) Visible(ts types.TS) bool {
+	return o.CreateTime.LessEq(&ts) &&
+		(o.DeleteTime.IsEmpty() || ts.Less(&o.DeleteTime))
+}
+
+func (o ObjectEntry) Location() objectio.Location {
+	return o.ObjectLocation()
+}
+
+func (o ObjectInfo) StatsValid() bool {
+	return o.ObjectStats.Rows() != 0
+}
+
+// sharedStates is shared among all PartitionStates
+type sharedStates struct {
+	sync.Mutex
+	// last block flush timestamp for table
+	lastFlushTimestamp types.TS
+}
+
+// RowEntry represents a version of a row
+type RowEntry struct {
+	BlockID types.Blockid // we need to iter by block id, so put it first to allow faster iteration
+	RowID   types.Rowid
+	Time    types.TS
+
+	ID                int64 // a unique version id, for primary index building and validating
+	Deleted           bool
+	Batch             *batch.Batch
+	Offset            int64
+	PrimaryIndexBytes []byte
+}
+
+func (r RowEntry) Less(than RowEntry) bool {
+	// asc
+	cmp := r.BlockID.Compare(than.BlockID)
+	if cmp < 0 {
+		return true
+	}
+	if cmp > 0 {
+		return false
+	}
+	// asc
+	if r.RowID.Less(than.RowID) {
+		return true
+	}
+	if than.RowID.Less(r.RowID) {
+		return false
+	}
+	// desc
+	if than.Time.Less(&r.Time) {
+		return true
+	}
+	if r.Time.Less(&than.Time) {
+		return false
+	}
+	return false
+}
+
+type PrimaryIndexEntry struct {
+	Bytes      []byte
+	RowEntryID int64
+
+	// fields for validating
+	BlockID types.Blockid
+	RowID   types.Rowid
+	Time    types.TS
+}
+
+func (p *PrimaryIndexEntry) Less(than *PrimaryIndexEntry) bool {
+	if res := bytes.Compare(p.Bytes, than.Bytes); res < 0 {
+		return true
+	} else if res > 0 {
+		return false
+	}
+	return p.RowEntryID < than.RowEntryID
+}
+
+type ObjectIndexByTSEntry struct {
+	Time         types.TS // insert or delete time
+	ShortObjName objectio.ObjectNameShort
+
+	IsDelete     bool
+	IsAppendable bool
+}
+
+func (b ObjectIndexByTSEntry) Less(than ObjectIndexByTSEntry) bool {
+	// asc
+	if b.Time.Less(&than.Time) {
+		return true
+	}
+	if than.Time.Less(&b.Time) {
+		return false
+	}
+
+	cmp := bytes.Compare(b.ShortObjName[:], than.ShortObjName[:])
+	if cmp < 0 {
+		return true
+	}
+	if cmp > 0 {
+		return false
+	}
+
+	//if b.IsDelete && !than.IsDelete {
+	//	return true
+	//}
+	//if !b.IsDelete && than.IsDelete {
+	//	return false
+	//}
+
+	return false
+}
+
+var nextRowEntryID = int64(1)
